requirements.txt
transformers==4.37.2
SpeechRecognition==3.10.1
torchaudio==2.2.0
torch==2.2.0
google-cloud-translate==3.15.1
fastapi==0.108.0
pygame==2.5.2
soundfile==0.12.1
numpy==1.26.4
pydub==0.25.1